requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_community
langchain_google_genai
langchain
streamlit
python-dotenv
pypdf